Working on the database abstraction
