+++ crates/cdk-cli/src/sub_commands/melt.rs
@@ -5,7 +5,8 @@ use std::str::FromStr;
 use anyhow::{bail, Result};
 use cdk::amount::MSAT_IN_SAT;
 use cdk::nuts::{CurrencyUnit, MeltOptions};
-use cdk::wallet::multi_mint_wallet::{MultiMintWallet, WalletKey};
+use cdk::wallet::multi_mint_wallet::MultiMintWallet;
+use cdk::wallet::types::WalletKey;
 use cdk::Bolt11Invoice;
 use clap::Args;
