+++ crates/cdk-cli/src/sub_commands/receive.rs
@@ -6,7 +6,8 @@ use anyhow::{anyhow, Result};
 use cdk::cdk_database::{self, WalletDatabase};
 use cdk::nuts::{SecretKey, Token};
 use cdk::util::unix_time;
-use cdk::wallet::multi_mint_wallet::{MultiMintWallet, WalletKey};
+use cdk::wallet::multi_mint_wallet::MultiMintWallet;
+use cdk::wallet::types::WalletKey;
 use cdk::wallet::Wallet;
 use cdk::Amount;
 use clap::Args;
