+++ crates/cdk-cli/src/sub_commands/send.rs
@@ -5,8 +5,7 @@ use std::str::FromStr;
 use anyhow::{bail, Result};
 use cdk::amount::SplitTarget;
 use cdk::nuts::{Conditions, CurrencyUnit, PublicKey, SpendingConditions};
-use cdk::wallet::multi_mint_wallet::WalletKey;
-use cdk::wallet::types::SendKind;
+use cdk::wallet::types::{SendKind, WalletKey};
 use cdk::wallet::MultiMintWallet;
 use cdk::Amount;
 use clap::Args;
