+++ crates/cdk-common/src/lightning.rs
