+++ crates/cdk-redb/Cargo.toml
@@ -7,20 +7,23 @@ description = "Redb storage backend for CDK"
 license = "MIT"
 homepage = "https://github.com/cashubtc/cdk"
 repository = "https://github.com/cashubtc/cdk.git"
-rust-version = "1.66.0" # MSRV
+rust-version = "1.66.0"                            # MSRV
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
 default = ["mint", "wallet"]
-mint = ["cdk/mint"]
-wallet = ["cdk/wallet"]
+mint = []
+wallet = []
 
 [dependencies]
 async-trait = "0.1"
-cdk = { path = "../cdk", version = "0.6.0", default-features = false }
+cdk-common = { path = "../cdk-common", version = "0.6.0" }
 redb = "2.1.0"
 thiserror = "1"
-tracing = { version = "0.1", default-features = false, features = ["attributes", "log"] }
+tracing = { version = "0.1", default-features = false, features = [
+    "attributes",
+    "log",
+] }
 serde = { version = "1", default-features = false, features = ["derive"] }
 serde_json = "1"
 lightning-invoice = { version = "0.32.0", features = ["serde", "std"] }
