+++ crates/cdk-sqlite/src/mint/error.rs
@@ -10,31 +10,31 @@ pub enum Error {
     SQLX(#[from] sqlx::Error),
     /// NUT00 Error
     #[error(transparent)]
-    CDKNUT00(#[from] cdk::nuts::nut00::Error),
+    CDKNUT00(#[from] cdk_common::nuts::nut00::Error),
     /// NUT01 Error
     #[error(transparent)]
-    CDKNUT01(#[from] cdk::nuts::nut01::Error),
+    CDKNUT01(#[from] cdk_common::nuts::nut01::Error),
     /// NUT02 Error
     #[error(transparent)]
-    CDKNUT02(#[from] cdk::nuts::nut02::Error),
+    CDKNUT02(#[from] cdk_common::nuts::nut02::Error),
     /// NUT04 Error
     #[error(transparent)]
-    CDKNUT04(#[from] cdk::nuts::nut04::Error),
+    CDKNUT04(#[from] cdk_common::nuts::nut04::Error),
     /// NUT05 Error
     #[error(transparent)]
-    CDKNUT05(#[from] cdk::nuts::nut05::Error),
+    CDKNUT05(#[from] cdk_common::nuts::nut05::Error),
     /// NUT07 Error
     #[error(transparent)]
-    CDKNUT07(#[from] cdk::nuts::nut07::Error),
+    CDKNUT07(#[from] cdk_common::nuts::nut07::Error),
     /// Secret Error
     #[error(transparent)]
-    CDKSECRET(#[from] cdk::secret::Error),
+    CDKSECRET(#[from] cdk_common::secret::Error),
     /// BIP32 Error
     #[error(transparent)]
     BIP32(#[from] bitcoin::bip32::Error),
     /// Mint Url Error
     #[error(transparent)]
-    MintUrl(#[from] cdk::mint_url::Error),
+    MintUrl(#[from] cdk_common::mint_url::Error),
     /// Could Not Initialize Database
     #[error("Could not initialize database")]
     CouldNotInitialize,
@@ -46,7 +46,7 @@ pub enum Error {
     Serde(#[from] serde_json::Error),
 }
 
-impl From<Error> for cdk::cdk_database::Error {
+impl From<Error> for cdk_common::database::Error {
     fn from(e: Error) -> Self {
         Self::Database(Box::new(e))
     }
